pyproject.toml
[project]
name = "kent-date-enough-hic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.72.0",
    "google-generativeai>=0.8.6",
    "kindling",
    "multipart>=1.3.1",
]

[tool.uv.sources]
kindling = { git = "https://github.com/Monster0506/kindling.git", rev = "ac0b98c" }
